Conflicts:
	Projects/TestBed/Classes/Infrastructure/GameCore.cpp

Former-commit-id: a54f8e68f7254876b896927cbd7cc7195b1d8823

+++ Sources/Internal/Render/2D/TextBlockGraphicRender.cpp
@@ -175,7 +175,7 @@ void TextBlockGraphicRender::Draw(const Color& textColor, const Vector2* offset)
     batch.texCoordPointer = vertexBuffer[0].texCoord.data;
     batch.textureSetHandle = graphicFont->GetTexture()->singleTextureSet;
     batch.samplerStateHandle = graphicFont->GetTexture()->samplerStateHandle;
-    batch.vertexCount = vertexBuffer.size();
+    batch.vertexCount = static_cast<uint32>(vertexBuffer.size());
     batch.indexPointer = indexBuffer;
     batch.indexCount = batch.vertexCount * 6 / 4;
     batch.worldMatrix = &offsetMatrix;
